requirements.txt
Django~=4.0.2
djangorestframework
